第三組程式/requirement.txt
absl-py==2.1.0
astunparse==1.6.3
attrs==23.2.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
contourpy==1.1.1
cycler==0.12.1
decorator==4.4.2
distlib==0.3.8
filelock==3.13.3
flatbuffers==24.3.25
fonttools==4.50.0
gast==0.4.0
google-auth==2.29.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.62.1
h5py==3.10.0
idna==3.6
imageio==2.35.1
imageio-ffmpeg==0.5.1
importlib_metadata==7.1.0
importlib_resources==6.4.0
imutils==0.5.4
jax==0.4.13
keras==2.13.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.4.5
KneronPLUS @ file:///C:/mypython4/counterV3/packages/KneronPLUS-2.0.0-py3-none-any.whl#sha256=dc087fd1fc3b937cd5677169f3053ddea6c9591f92a61a5e3aba5c0c0eb7dea2
libclang==18.1.1
Markdown==3.6
MarkupSafe==2.1.5
matplotlib==3.7.5
mediapipe==0.10.11
ml-dtypes==0.2.0
moviepy==1.0.3
mysql-connector-python==9.0.0
numpy==1.23.3
oauthlib==3.2.2
opencv-contrib-python==4.9.0.80
opencv-python==4.6.0.66
opt-einsum==3.3.0
packaging==24.0
pillow==10.2.0
platformdirs==4.2.0
playsound==1.3.0
proglog==0.1.10
protobuf==3.20.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.21
pygame==2.6.1
pyparsing==3.1.2
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
python-dateutil==2.9.0.post0
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
scipy==1.9.3
six==1.16.0
sounddevice==0.4.6
tensorboard==2.13.0
tensorboard-data-server==0.7.2
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-intel==2.13.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
tqdm==4.67.0
typing_extensions==4.5.0
urllib3==2.2.1
virtualenv==20.25.1
Werkzeug==3.0.1
wrapt==1.16.0
zipp==3.18.1
